feat #13: ReviewStar 별도 파일로 분리
